+++ src/app/(root)/(routes)/login/page.tsx
@@ -1,4 +1,4 @@
-import LogInForm from '@/components/organisms/LogInForm'
+import { LogInForm } from '@/components/organisms/LogInForm'
 
 export default function Login() {
   return <LogInForm />
