+++ src/components/atoms/Button/index.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import Button from '.'
+import { Button } from '.'
 
 const meta = {
   title: 'Component/Button',
