+++ src/components/atoms/Button/index.tsx
@@ -1,32 +1,3 @@
-'use client'
+import Button from './Button'
 
-import './index.scss'
-
-type ButtonProps = {
-  text?: string
-  color?: 'yellow' | 'red'
-  width?: number
-  height?: number
-  onClick(): void
-}
-
-export default function Button({
-  text,
-  color,
-  width,
-  height,
-  onClick,
-}: ButtonProps) {
-  return (
-    <button
-      onClick={onClick}
-      className={`button-container ${color}`}
-      style={{
-        width: `${width}rem`,
-        height: `${height}rem`,
-      }}
-    >
-      {text}
-    </button>
-  )
-}
+export { Button }
