+++ src/components/molcules/SignTitle/index.tsx
@@ -1,27 +1,3 @@
-'use client'
+import SignTitle from './SignTitle'
 
-import Image from 'next/image'
-import Assets from '@/config/assets'
-import './index.scss'
-
-interface SignTitleProps {
-  text?: string
-}
-
-const SignTitle = ({ text }: SignTitleProps) => {
-  return (
-    <>
-      <div className="sign-title">
-        <Image
-          src={Assets.TITLELOGO_SVG_PATH}
-          width={50}
-          height={50}
-          alt="titleLogo"
-        />
-        {text}
-      </div>
-    </>
-  )
-}
-
-export default SignTitle
+export { SignTitle }
